refactor(docuflow): decompose 930-line editor page
The monolith becomes a 48-line orchestrator plus eight focused
subcomponents and a data module. Layout:

  DocuflowBackground.vue  — fixed grid + radial gradient
  EditorChrome.vue        — dots + tabs + meta pills
  SourcePane.vue          — textarea + line-numbers + cursor tracking
  PreviewPane.vue         — marked -> v-html with the :deep() markdown styles
  EditorSection.vue       — composes chrome/source/divider/preview/statusbar;
                            owns splitPct drag state and cursor state
  StepsSection.vue        — the "write, preview, ship" 3-step section
  ShotsSection.vue        — dark-mode screenshots section
  TechSection.vue         — tech grid + GitHub CTA

The rendered-markdown :deep() rules (h1-h3, p, blockquote, code, pre,
a, ul/ol, hr, strong, em) stay colocated with PreviewPane so they
survive component extraction. The resizable split-pane JS (mousedown
-> window listeners) stays in EditorSection where splitPct lives.

No visual change: 'bun run test:visual -- --grep docuflow' green on
mobile/tablet/desktop.
